Add: safe signed unsigned int comparsions

+++ daemon/util/FileSystem.cpp
@@ -745,8 +745,8 @@ std::optional<FileSystem::DiskState> FileSystem::GetDiskState(const char* path)
 
 	if (GetDiskFreeSpaceEx(path, &freeBytesAvailable, &totalNumberOfBytes, nullptr))
 	{
-		size_t available = static_cast<size_t>(freeBytesAvailable.QuadPart);
-		size_t total = static_cast<size_t>(totalNumberOfBytes.QuadPart);
+		size_t available = freeBytesAvailable.QuadPart;
+		size_t total = totalNumberOfBytes.QuadPart;
 		return FileSystem::DiskState{ available, total };
 	}
 #else
